+++ src/icemac/addressbook/preferences/tests/test_frontend.py
@@ -4,18 +4,14 @@ import pytest
 @pytest.mark.webdriver
 def test_preferences__1(address_book, webdriver):
     """Testing JS funtions of preferences frontend."""
-    sel = webdriver.login('mgr')
-    sel.open("/ab/++preferences++/ab")
+    prefs = webdriver.prefs
+    webdriver.login('mgr', prefs.PREFS_URL)
     # Preference groups are closed by default, fields are not visible:
-    sel.assertNotVisible("css=#form-widgets-columns-row")
-    # After opening the group the field is shown:
-    sel.click("//fieldset[@class='personLists']/legend")
-    sel.waitForVisible("css=#form-widgets-columns-row")
+    prefs.wait_for_fields_visible(False)
+    prefs.toggle_group('personLists')
+    prefs.wait_for_fields_visible(True)
     # Clicking on a form element does not close the group:
-    sel.addSelection("id=form-widgets-columns-from",
-                     "label=person -- birth date")
-    sel.click("name=from2toButton")
-    sel.assertVisible("css=#form-widgets-columns-row")
-    # Clicking on the legend closes the group:
-    sel.click("//fieldset[@class='personLists']/legend")
-    sel.waitForNotVisible("css=#form-widgets-columns-row")
+    prefs.select_column('person -- birth date')
+    prefs.wait_for_fields_visible(True)
+    prefs.toggle_group('personLists')
+    prefs.wait_for_fields_visible(False)